A preset based on my personal default settings
I've gotten pretty used to these settings over the past decade or so.
It's maybe a bit weird to put the git branch/status on the left,
so it shifts the current directory to the right when it's visible,
but I like that it makes it more visually obvious when I'm in a git repo.
Bright yellow isn't everyone's favorite color, but using it for the
current directory makes a very clear visual divider between each command,
since commands rarely use bright yellow in their output.