Fix accidentally affecting global settings
Putting the 'setopt no_local_options prompt_percent' at the end of yazpt_precmd was causing the changes made by 'emulate -L zsh' to persist after the function exited. I noticed because auto_cd wasn't set anymore, and trying to turn it on didn't seem to work (because it got turned off again as the next prompt was generated).

Could rename the current yazpt_precmd to like yazpt_precmd_impl, and make a new tiny yazpt_precmd which just calls yazpt_precmd_impl and then forces prompt_percent on, but that seems overkill, and forcing the setting on was a questionable decision anyway. Let's just turn it on as yazpt.zsh-theme is sourced, and call it good.