Detect Path Finder / TermKit on macOS
I’m guessing Path Finder uses https://swiftpackageindex.com/migueldeicaza/SwiftTerm, which is part of https://github.com/migueldeicaza/TermKit, but I don’t know for sure.

It would be better if we could check for the $TERMKIT_HOST_APP environment variable down in the “?6” block with Haiku Terminal, since that’s how Path Finder usually behaves, but during app startup, it doesn’t answer the terminal query at all, so it gets detected as “unknown” if we do it that way.