Wait longer for a response from the terminal about cursor position
Sometimes when terminal emulators are starting up, they don't respond to the "blank" segment's escape-sequence query for the cursor position quickly enough to avoid hitting our 0.1 second timeout threshold, and the latter part of the response sequence ends up getting splatted onto the command line as though it was typed. Sometimes just the "R" at the end, sometimes more.

I've seen it various places, but can repro it most reliably on macOS by leaving a couple Apple Terminal windows open, and restarting with "Reopen windows when logging back in" checked.

Setting a longer timeout fixes the problem, or at least makes it far less likely to happen. But out of worry that the longer timeout will make yazpt unusably slow if a terminal emulator doesn't respond to our cursor position query, now if we ever can't get a response within our timeout period, we don't ask for cursor position again in that terminal session. Not sure how useful that is, because (a) I haven't been able to find a terminal emulator that doesn't respond to the query, to do real-life testing with, (b) the initial 5-second delay is pretty painful anyway, and (c) if the terminal emulator responds at all, but with a sequence that doesn't end with "R", zsh's read command will hang forever anyway. (That last one has always been true, I just didn't realize it until now.)

But this is definitely an improvement over the status quo.